Implement basic login template
Added account and response models and basic login/register logic. I had to manually implement getters/setters for the models as
lombok wouldn't work for me. I also deleted the interface for the repository and added delete and get functionalitites for the
account controller. I also did some renaming and cleaning up of the code.