Implement forgot password code creation and reset process
Implemented logic to create and store reset codes in database. Each reset code has a lifespan of 10
minutes and there can only be one reset code per account. You can only "guess" the code for an
account three times until it gets deleted. The code exists of 8 random digits and gets sent to
the email of the user. The user has to send the own email as we only save hashed passwords. If
the hashes match, we create a new reset code.